refactor: introduce `snow/validators/statetest`
